replaced the previous with new
add all the new files( i added middleware as well as authentication backends)
