--- rest/restapp/views.py
@@ -1,19 +1,0 @@
-from django.contrib.auth.models import User
-from .serializers import RegisterSerializer
-from rest_framework import generics
-from rest_framework.permissions import AllowAny
-from allauth.socialaccount.providers.google.views import GoogleOAuth2Adapter
-from allauth.socialaccount.providers.oauth2.client import OAuth2Client
-from dj_rest_auth.registration.views import SocialLoginView
-
-
-class RegisterView(generics.CreateAPIView):
-    queryset = User.objects.all()
-    permission_classes = (AllowAny,)
-    serializer_class = RegisterSerializer
-
-
-class GoogleLogin(SocialLoginView):
-    adapter_class = GoogleOAuth2Adapter
-    callback_url = "http://localhost:3000/api/auth/callback/google"
-    client_class = OAuth2Client
